+++ app/ide-desktop/lib/content/src/index.ts
@@ -8,7 +8,7 @@ import * as authentication from 'enso-authentication'
 import * as common from 'enso-common'
 import * as contentConfig from 'enso-content-config'
 
-import * as app from '../../../../../target/ensogl-pack/linked-dist/index'
+import * as app from '../../../../../target/ensogl-pack/linked-dist'
 import GLOBAL_CONFIG from '../../../../gui/config.yaml' assert { type: 'yaml' }
 
 const logger = app.log.logger