+++ app/ide-desktop/lib/dashboard/src/authentication/src/dashboard/localBackend.ts
@@ -46,6 +46,14 @@ export class LocalBackend implements Partial<backend.Backend> {
             modifiedAt: project.lastOpened,
             parentId: newtype.asNewtype<backend.AssetId>(''),
             permissions: [],
+            projectState: {
+                type:
+                    project.id === LocalBackend.currentlyOpeningProjectId
+                        ? backend.ProjectState.openInProgress
+                        : project.lastOpened != null
+                        ? backend.ProjectState.closed
+                        : backend.ProjectState.created,
+            },
         }))
     }