+++ app/ide-desktop/lib/types/globals.d.ts
@@ -1,6 +1,9 @@
 /** @file Globals defined outside of TypeScript files.
  * These are from variables defined at build time, environment variables,
  * monkeypatching on `window` and generated code. */
+// This file is being imported for its types.
+// eslint-disable-next-line no-restricted-syntax
+import * as buildJson from './build.json' assert { type: 'json' }
 
 // =============
 // === Types ===
@@ -14,14 +17,6 @@ interface StringConfig {
 /** The public interface exposed to `window` by the IDE. */
 interface Enso {
     main: (inputConfig?: StringConfig) => Promise<void>
-}
-
-/** Build information injected by the build script. */
-interface BuildInfo {
-    commit: string
-    version: string
-    engineVersion: string
-    name: string
 }
 
 // ==========================
@@ -72,7 +67,7 @@ declare global {
     // These are used in other files (because they're globals)
     /* eslint-disable @typescript-eslint/naming-convention */
     const BUNDLED_ENGINE_VERSION: string
-    const BUILD_INFO: BuildInfo
+    const BUILD_INFO: buildJson.BuildInfo
     const PROJECT_MANAGER_IN_BUNDLE_PATH: string
     const IS_DEV_MODE: boolean
     // This will be `undefined` when it is not defined by esbuild.
@@ -80,6 +75,3 @@ declare global {
     const REDIRECT_OVERRIDE: string | undefined
     /* eslint-disable @typescript-eslint/naming-convention */
 }
-
-// Required for the globals to be visible to other files.
-export {}